+++ src/core/errors.ts
@@ -1,0 +1,7 @@
+export class CancelledError extends Error {
+  constructor(message = 'Operation canceled.') {
+    super(message)
+    this.name = 'CancelledError'
+  }
+}
+